Fix bundler

+++ utils/bundle.mjs
@@ -3,6 +3,9 @@ import fs from "fs-extra";
 
 fs.removeSync("./dist");
 
-fs.copySync("./public", "./dist/public");
+// Check if ./public exists
+if (fs.existsSync("./public")) {
+  fs.copySync("./public", "./dist/public");
+}
 fs.copySync("./.next/standalone", "./dist");
 fs.copySync("./.next/static", "./dist/.next/static");
